requirements.txt
pre-commit==4.2.*
ipywidgets==8.1.*
jsonargparse[signatures,omegaconf]==4.38.*
matplotlib==3.10.*
torchinfo==1.8.*
datasets==3.6.*
lightning==2.5.*
tensorboard==2.19.*
